pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watchtower"
version = "0.1.0"
description = "Vagt - Shift Management System for FynBus"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Gorm Reventlow"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "Django>=5.0,<6.0",
    "python-decouple>=3.8",
    "django-htmx>=1.17",
    "djangorestframework>=3.14",
    "whitenoise>=6.6",
    "gunicorn>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "pre-commit>=3.5",
    "django-debug-toolbar>=4.2",
]

[project.urls]
Homepage = "https://github.com/gorm/watchtower"
Repository = "https://github.com/gorm/watchtower"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "config*"]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
